feat: moves fetchVolunteers to lib
