+++ src/lib/index.ts
@@ -1,0 +1,1 @@
+export { default as fetchVolunteers } from "./fetchVolunteers";
