+++ src/utils/index.ts
@@ -10,5 +10,4 @@ export { default as encrypt } from "./encrypt";
 export { default as getFirstName } from "./getFirstName";
 export { default as saveBusaraABExperiment } from "./saveBusaraABExperiment";
 export { default as notFoundErrorPayload } from "./notFoundErrorPayload";
-export { default as fetchVolunteers } from "./fetchVolunteers";
 export * from "./stringfyBigInt";
